feat: User profile page without posts (#22)
A user profile page is added, to show the basic info of a user - name, bio, and preferred art mediums. In the header, the profile link was temporarily hardcoded to point to an existing profile.

On the backend, a GET /api/user/:name endpoint was set up. As this was the first time the database was used, the MongoDB configuration was added as well.

---------

Co-authored-by: Coralee <45268389+Coralee-Long@users.noreply.github.com>